pull calculate age into own util
