+++ src/utils/age-utils.ts
@@ -1,0 +1,15 @@
+function calculateAge(birthday: string) {
+  const birthdate = new Date(birthday);
+  const today = new Date();
+  let age = today.getFullYear() - birthdate.getFullYear();
+  const monthDiff = today.getMonth() - birthdate.getMonth();
+  if (
+    monthDiff < 0 ||
+    (monthDiff === 0 && today.getDate() < birthdate.getDate())
+  ) {
+    age--;
+  }
+  return age;
+}
+
+export { calculateAge };
